Create Correct Smartmaker Smarthost 16X2 Board Pin Config
The Smartmaker Smarthost LCD 16X2 boards (Smarthost LCD PRO, Smarthost LCD 16X2 MOD, and Smarthost LCD 16X2) are all carrier boards for Hitachi HD44780 driver-compatible 16x2 LCD displays.

In theory, these boards should "just work" with the Arduino LiquidCrystal library. The examples for that lib are all built around this de facto standard interface. And according to the document in the Smartmaker Wiki, the pins used in the example code (digital pins 12, 11, 5, 4, 3, and 2) are supposed to be the same one that Smarthost LCD boards use: Smartbus pins B13, B12, B06, B05, B04, and B03, which map to the Arduino pins in question.

The correct pins

It turns out the reality of the boards doesn't match the Smartmaker wiki documentation. The the upper values pins (11/B12 and 12/B13) are not actually connected at all. You can see this pretty easily on the MOD, where no traces connect to those pins on the board!

This table has the correct mapping.